Update README.md
a technical context option (listing all known elements, assets, and functions and otherwise available concepts, etc. within digital relevance - related terms or functions for programming languages, any executable code, for example.) of course, a flip side option for antonyms - should probably be just as developed and advanced as the synonym generator tbh. i suppose it’ll be both ! and as for the technical feature; we’ll particularly make this as advanced as we possibly can - could be very useful for opposing/reversing/opposite/undo/reverse engineering usage cases. a creative writing symbolic feature that can be very useful and powerful for any writer or author for literature, novelty, and creative work/art honestly we could probably apply every professional field as an optional specification feature. medical, dental, environmental, research, legal, entrepreneurial, social, etc. this one would be a great innovation: finish each other’s sandwiches feature. come up with the word they’re forgetting or can’t think of. the idea is defined in their mind and the point that they are trying to make is an idea they want help with putting it together into a coherent sentence. or maybe they don’t even have the concept of what they’re trying to accomplish, it might be in an abstract form of a notion or a feeling of something they know exists in a transcendental next undiscovered breakthrough that has yet to be interpreted and articulated as a tangible basis to apply and theorize into effective integration to expand our intellectual domain. maybe they’re just getting writer’s block from this one thing. or maybe they actually are stuck because of a hardcore case of writer’s block. i mean, no matter which way you put it - giving them the right word to use or finishing their sentence (they might even finish it themselves after back and forth bouncing off each other’s banters to investigate the situation. they were able to conclusively solve the case with this help. possible ways to make this truly helpful, effectively useful, and even interestingly intelligent would be to start from a very basic foundation - perhaps putting just whatever context they have so far for the sentence or maybe they just have a single word in mind. or they can just explain it as best they can “what’s that word, you know for when people do things like this lol. a universal human experience for sure. i’ve had this scenario play out countless times. it’s a canon event. a recurring filler, if you will. and i think i came up with the word i was looking for to wrap this up! inception.” ok so this happens to people everyday whether they keep it to themselves or they request assistance externally.